requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
pillow
pandas
numpy
scikit-learn
requests